[#86] refactor(payment): 결제 도메인의 ENUM을 common으로 이동
