+++ service/payment/payment-core/src/main/java/radiata/service/payment/core/implementation/PaymentSaver.java
@@ -5,9 +5,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
 import radiata.common.annotation.Implementation;
+import radiata.common.domain.payment.constant.PaymentType;
 import radiata.service.payment.core.domain.model.entity.Payment;
 import radiata.service.payment.core.domain.model.vo.Money;
-import radiata.service.payment.core.domain.model.vo.PaymentType;
 import radiata.service.payment.core.domain.repository.PaymentRepository;
 
 @Slf4j
