+++ service/payment/payment-core/src/main/java/radiata/service/payment/core/service/TossPaymentService.java
@@ -4,10 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import radiata.common.domain.payment.constant.PaymentStatus;
 import radiata.common.domain.payment.dto.request.TossPaymentCreateRequestDto;
 import radiata.common.domain.payment.dto.response.TossPaymentCreateResponseDto;
 import radiata.service.payment.core.domain.model.entity.Payment;
-import radiata.service.payment.core.domain.model.vo.PaymentStatus;
 import radiata.service.payment.core.implementation.PaymentRequester;
 import radiata.service.payment.core.implementation.PaymentSaver;
