+++ service/payment/payment-core/src/test/java/radiata/service/payment/core/domain/model/vo/BankTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import radiata.common.domain.payment.constant.Bank;
 
 @DisplayName("Bank VO 테스트")
 class BankTest {
